aggiunto parseFloat

+++ js/script.js
@@ -1,5 +1,5 @@
 
-let kmUser = parseInt( prompt('Inserire il totale dei km da percorrere'));
+let kmUser = parseFloat( prompt('Inserire il totale dei km da percorrere'));
 let agePassenger = parseInt( prompt("Inserire l'età del passeggero"));
 
 let standardPrice = kmUser * 0.21;
